TST (string dtype): duplicate pandas/tests/indexes/object tests specifically for string dtypes
